Renamed read and write test files
