+++ src/read.c
@@ -7,7 +7,7 @@
 #include <errno.h> // For errno
 #include <stdlib.h> // For malloc
 #include <time.h> // for clock_gettime & timespec struct
-#include "write_test.h"
+#include "read.h"
 
 #define LOG(...) \
 	if(verbose_flag == 1){ \
@@ -24,14 +24,11 @@
 		return errno; \
 	}
 
-void fill_with_random(char* buffer, size_t size);
-
-int write_test(
+int read_test(
 	int fd,
 	size_t chunk_size,
 	size_t chunks,
 	int verbose_flag,
-	int random_flag,
 	double *result
 ){
 	char* chunk;
@@ -40,21 +37,16 @@ int write_test(
 	ERR_CHECK(chunk == NULL, "malloc failed");
 	LOG("Allocated %d bytes chunk\n", chunk_size);
 
-	if(random_flag == 1){
-		fill_with_random(chunk, chunk_size);
-		LOG("Filled chunk with random bytes\n");
-	}
-	
 	struct timespec start;
 	struct timespec end;
 	int start_result, end_result;
 
 	start_result = clock_gettime(CLOCK_REALTIME, &start); // Start time
 
-	int wr;
+	int rr;
 	for(size_t n = 0; n < chunks; n++){
-		wr = write(fd, chunk, chunk_size);
-		ERR_CHECK(wr < 1, "Failed to write"); //TODO Check if it should be `< 1` or `< 0`
+		rr = read(fd, chunk, chunk_size);
+		ERR_CHECK(rr < 1, "Failed to read");  //TODO Check if it should be `< 1` or `< 0`
 	}
 
 	end_result = clock_gettime(CLOCK_REALTIME, &end);
@@ -70,10 +62,4 @@ int write_test(
 	LOG("Freed %d bytes chunk\n", chunk_size);
 
 	return 0;
-}
-
-void fill_with_random(char* buffer, size_t size){
-	for(size_t n = 0; n < size; n++){
-		buffer[n] = (char)rand();
-	}
 }
