+++ src/write.h
